requirements.txt
numpy
opencv-python
pyrender
trimesh
open3d
rembg
torch
diffusers
transformers
Pillow
matplotlib
